Move versions to separate module, low-level messaging to Peer
