+++ tests/ClientProtocol.hs
@@ -9,8 +9,7 @@ import qualified Hedgehog.Range as Range
 import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.Hedgehog (testProperty)
 
-import qualified MagicWormhole.Internal.ClientProtocol as ClientProtocol
-import qualified MagicWormhole.Internal.Pake as Pake
+import qualified MagicWormhole.Internal.Peer as ClientProtocol
 
 
 tests :: IO TestTree
@@ -18,7 +17,7 @@ tests = pure $ testGroup "Peer"
   [ testProperty "SecretBox encryption roundtrips" $ property $ do
       purpose <- forAll $ Gen.bytes (Range.linear 0 10)
       secret <- forAll $ Gen.bytes (Range.linear 0 10)
-      let key = ClientProtocol.deriveKey (Pake.SessionKey secret) purpose
+      let key = ClientProtocol.deriveKey (ClientProtocol.SessionKey secret) purpose
       plaintext <- forAll $ Gen.bytes (Range.linear 1 256)
       ciphertext <- liftIO $ ClientProtocol.encrypt key plaintext
       let decrypted = ClientProtocol.decrypt key ciphertext
